+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## 0.7.0
 
+- Added Mailing Support when jobs have completed
 - Remove support for sqlite migration (SL System 4.0.0 to SL System 5.0.0)